refactor(types): move game object node type
- Extract `GameObjectNode` into a dedicated shared type module
- Update render and context helpers to import the node type from `src/types`
- Remove the node type definition from scene helpers and re-export it through
  the types index
