+++ src/helpers/scene.ts
@@ -9,12 +9,3 @@ export function getScene<Type = Scene>(): Type {
 export function setScene(scene: Scene): void {
   _scene = scene;
 }
-
-/**
- * Info about a tracked game object in the tree.
- */
-export interface GameObjectNode {
-  gameObject: Phaser.GameObjects.GameObject;
-  props: Record<string, unknown>;
-  children: (GameObjectNode | null)[];
-}
