+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './events';
+export * from './node';
 export * from './phaser';
 export * from './props';
 export * from './utils';
